algofun_service/requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
cryptography==41.0.3
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-filter==23.2
djangorestframework==3.14.0
djwto==0.0.3
Faker==19.3.1
flake8==4.0.1
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.3
Markdown==3.4.4
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
promise==2.3
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.9
whitenoise==5.3.0
django-polymorphic==3.1.0
